Add support for prometheus metrics
- Add support for prometheus metrics
- Split server to multiple files for readability
