+++ requirements.txt
@@ -1,2 +1,3 @@
 httpx==0.28.1
 fastmcp==2.8.0
+prometheus_client==0.20.0
